Flyttet SecurityFilterChain til ApplicationConfig
